fix: align Helm lifecycle history semantics
